third_party/WebKit/LayoutTests/bluetooth/requestDevice/blocklisted-service-in-filter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('BlocklistTestAdapter')
    .then(() => assert_promise_rejects_with_message(
      requestDeviceWithKeyDown({
        filters: [{services: ['human_interface_device']}]}),
      new DOMException(
        'requestDevice() called with a filter containing a blocklisted UUID. ' +
        'https://goo.gl/4NeimX',
        'SecurityError'),
      'Requesting blocklisted service rejects.'));
}, 'Reject with SecurityError if requesting a blocklisted service.');
</script>
